确认删除按钮

+++ src/view/article/home.vue
@@ -149,9 +149,19 @@ export default {
     },
 
     delete ({ id, index }) {
-      deleteArticle(id).then(res => {
-        this.$Message.success('删除成功')
-        this.dataSet.data = _.reject(this.dataSet.data, { id: id })
+      let title = this.dataSet.data[index].title
+
+      this.$Modal.confirm({
+        title: `删除${title}`,
+        content: '确认要删除这篇文章吗？',
+        loading: true,
+        onOk: () => {
+          deleteArticle(id).then(res => {
+            this.$Modal.remove()
+            this.$Message.success('删除成功')
+            this.dataSet.data = _.reject(this.dataSet.data, { id: id })
+          })
+        }
       })
     },
